+++ libft/ft_strlen.c
@@ -11,7 +11,7 @@
 /* ************************************************************************** */
 #include "libft.h"
 
-unsigned int	ft_strlen(char *s)
+size_t	ft_strlen(char *s)
 {
 	int	c;
