+++ libft/ft_strncmp.c
@@ -11,9 +11,9 @@
 /* ************************************************************************** */
 #include "libft.h"
 
-int	ft_strncmp(char *s1, char *s2, unsigned int n)
+int	ft_strncmp(char *s1, char *s2, size_t n)
 {
-	unsigned int	c;
+	size_t	c;
 
 	c = 0;
 	while ((c < n) && (s1[c] || s2[c]))
