+++ libft/libft.h
@@ -12,19 +12,20 @@
 
 #ifndef LIBC_H
 #define LIBC_H
+	#include <stdlib.h>
 
-	void			ft_bzero(void *s, unsigned int n);
+	void			ft_bzero(void *s, size_t n);
 	int				ft_isalnum(int c);
 	int				ft_isalpha(int c);
 	int				ft_isascii(int c);
 	int				ft_isdigit(int c);
 	int				ft_isprint(int c);
-	void			*ft_memcpy(void *dest, const void *src, unsigned int n);
-	void			*ft_memset(void *s, int c, unsigned int n);
-	unsigned int	ft_strlcat(char *dest, char *src, unsigned int size);
-	unsigned int	ft_strlcpy(char *dest, char *src, unsigned int size);
-	unsigned int	ft_strlen(char *s);
-	int				ft_strncmp(char *s1, char *s2, unsigned int n);
+	void			*ft_memcpy(void *dest, const void *src, size_t n);
+	void			*ft_memset(void *s, int c, size_t n);
+	size_t	ft_strlcat(char *dest, char *src, size_t size);
+	size_t	ft_strlcpy(char *dest, char *src, size_t size);
+	size_t	ft_strlen(char *s);
+	int				ft_strncmp(char *s1, char *s2, size_t n);
 	int				ft_toupper(int c);
 	int				ft_tolower(int c);
